projects/lnmarkets-funding-bot/requirements.txt
lnmarkets-sdk>=3.0.0
python-dotenv>=1.0.0
